kt_5/requirements.txt
allure-pytest==2.8.18
pytest==6.2.5
requests==2.24.0
structlog
logging==0.4.9.6
assertpy
jsonschema
pydantic